ticket:1542 - check if rasdaman is properly started and stop it if it is not.
Summary:
when rasservers fail to initialize properly, rasmgr will keep trying to restart them (every 3 seconds by default), unnecessarily producing many log files.
To fix this start_rasdaman.sh will try a simple select query after starting rasservers; if it fails to execute it means that no rasserver has probably started (select version() should always execute fine otherwise).
In this case it shuts down rasmgr and hence prevents it going into the above loop of constantly restarting the rasservers.

Test Plan: I've tested this manually by inserting a segfault into rasserver's initialization. It would be very hard to automate a test for this.

Reviewers: vmerticariu, bphamhuu

Reviewed By: vmerticariu, bphamhuu

Differential Revision: http://codereview.rasdaman.org/D385

+++ rasmgr_x/src/serverrasnet.cc
@@ -84,7 +84,10 @@ using rasnet::service::RasServerService;
 
 #define RASEXECUTABLE BINDIR"rasserver"
 
-const boost::uint32_t ServerRasNet::SERVER_CLEANUP_TIMEOUT = 30000;
+// give 3 seconds to rasserver to cleanly shutdown, before killing it with a SIGKILL
+const boost::int32_t ServerRasNet::SERVER_CLEANUP_TIMEOUT = 3000000;
+// 10 milliseconds
+const boost::int32_t ServerRasNet::SERVER_CHECK_INTERVAL = 10000;
 
 ServerRasNet::ServerRasNet(const ServerConfig& config)
 {
@@ -393,11 +396,23 @@ void ServerRasNet::stop(KillLevel level)
             LERROR << "Failed to send SIGTERM to server with ID:" << this->serverId;
         }
 
-        usleep(SERVER_CLEANUP_TIMEOUT);
-
-        if (kill(this->processId, SIGKILL))
+        // wait until the server process is dead
+        boost::int32_t cleanupTimeout = SERVER_CLEANUP_TIMEOUT;
+        bool isProcessAlive = true;
+        while (cleanupTimeout > 0 && isProcessAlive)
         {
-            LERROR << "Failed to send SIGKILL to server with ID:" << this->serverId;
+            usleep(SERVER_CHECK_INTERVAL);
+            cleanupTimeout -= SERVER_CHECK_INTERVAL;
+            isProcessAlive = (kill(this->processId, 0) == 0);
+        }
+
+        // if the server is still alive after SERVER_CLEANUP_TIMEOUT, send a SIGKILL
+        if (isProcessAlive)
+        {
+            if (kill(this->processId, SIGKILL))
+            {
+                LERROR << "Failed to send SIGKILL to server with ID:" << this->serverId;
+            }
         }
     }
     break;
